generate space type from schema
