+++ src/lib/app/entities.ts
@@ -4,7 +4,7 @@ import {AccountSchema} from '$lib/vocab/account/account';
 import {PersonaSchema} from '$lib/vocab/persona/persona';
 import {CommunitySchema} from '$lib/vocab/community/community';
 import {MembershipSchema} from '$lib/vocab/membership/membership';
-import {SpaceSchema} from '$lib/vocab/space/space';
+import {SpaceSchema} from '$lib/vocab/space/space.schema';
 import {EntitySchema} from '$lib/vocab/entity/entity';
 
 // TODO generate this file?
