+++ src/lib/vocab/space/space.schema.ts
@@ -1,0 +1,19 @@
+export const SpaceSchema = {
+	$id: '/schemas/Space.json',
+	type: 'object',
+	properties: {
+		space_id: {type: 'number'},
+		name: {type: 'string'},
+		url: {type: 'string'},
+		view: {
+			type: 'object',
+			tsType: 'ViewData',
+			tsImport: "import type {ViewData} from '$lib/vocab/view/view'",
+		},
+		created: {type: 'object', format: 'date-time', tsType: 'Date'},
+		updated: {type: ['object', 'null'], format: 'date-time', tsType: 'Date | null'},
+		community_id: {type: 'number'},
+	},
+	required: ['space_id', 'name', 'url', 'view', 'created', 'updated', 'community_id'],
+	additionalProperties: false,
+};
