+++ src/lib/vocab/space/space.ts
@@ -1,4 +1,6 @@
-import type {ViewData} from '$lib/vocab/view/view';
+// generated by src/lib/vocab/space/space.schema.ts
+
+import {type ViewData} from '$lib/vocab/view/view';
 
 export interface Space {
 	space_id: number;
@@ -9,18 +11,5 @@ export interface Space {
 	updated: Date | null;
 	community_id: number;
 }
-export const SpaceSchema = {
-	$id: '/schemas/Space.json',
-	type: 'object',
-	properties: {
-		space_id: {type: 'number'},
-		name: {type: 'string'},
-		url: {type: 'string'},
-		view: {type: 'object', tsType: 'ViewData'},
-		created: {type: 'object', format: 'date-time', tsType: 'Date'},
-		updated: {type: ['object', 'null'], format: 'date-time', tsType: 'Date | null'},
-		community_id: {type: 'number'},
-	},
-	required: ['space_id', 'name', 'url', 'view', 'created', 'updated', 'community_id'],
-	additionalProperties: false,
-};
+
+// generated by src/lib/vocab/space/space.schema.ts
